[skip-travis] non-working initial proposal
