+++ src/main/scala/com/johnsnowlabs/storage/RocksDBReader.scala
@@ -1,58 +1,66 @@
 package com.johnsnowlabs.storage
 
-import java.io.File
+import java.nio.{ByteBuffer, ByteOrder}
 
 import com.johnsnowlabs.nlp.util.LruMap
-import org.apache.spark.SparkFiles
-import org.rocksdb.RocksDB
+import spire.ClassTag
 
-/*
-  1. Copy Embeddings to local tmp file
-  2. Index Embeddings if need
-  3. Copy Index to cluster
-  4. Open RocksDb based Embeddings on local index (lazy)
- */
-abstract class RocksDBReader[A](val fileName: String, val caseSensitive: Boolean, lruCacheSize: Int = 100000)
-  extends RocksDBConnection with AutoCloseable {
+abstract class RocksDBReader[A: ClassTag](
+                                 connection: RocksDBConnection,
+                                 caseSensitiveIndex: Boolean,
+                                 lruCacheSize: Int = 100000) extends AutoCloseable {
+
+  connection.connectReadOnly
 
   @transient val lru = new LruMap[String, Option[Array[A]]](lruCacheSize)
 
-  lazy protected val localPath: String = StorageHelper.getLocalPath(fileName)
+  protected val emptyValue: A
 
-  def findLocalDb: RocksDB = {
-    if (isOpen)
-      getDb
-    else if (new File(localPath).exists()) {
-      openReadOnly(localPath)
+  protected def getFromBuffer(buffer: ByteBuffer, index: Int): A // wrapper.getFloat(...)
+
+  def fromBytes(source: Array[Byte]): Array[A] = {
+    val wrapper = ByteBuffer.wrap(source)
+    wrapper.order(ByteOrder.LITTLE_ENDIAN)
+    val result = Array.fill[A](source.length / 4)(emptyValue)
+
+    for (i <- result.indices) {
+      result(i) = getFromBuffer(wrapper, i * 4)
     }
-    else {
-      val localFromClusterPath = SparkFiles.get(fileName)
-      require(new File(localFromClusterPath).exists(), s"Storage not found under given ref: $fileName\n" +
-        s" This usually means:\n1. You have not loaded any storage under such ref\n2." +
-        s" You are trying to use cluster mode without a proper shared filesystem.\n3. source was not provided to Storage creation" +
-        s"\n4. If you are trying to utilize Storage defined elsewhere, make sure you it's appropriate ref. ")
-      openReadOnly(localFromClusterPath)
-    }
+    result
   }
 
-  protected def customizedLookup(index: String): Option[Array[A]]
+  private def lookupByIndex(index: String): Option[Array[A]] = {
+    lazy val resultLower = connection.getDb.get(index.trim.toLowerCase.getBytes())
+    lazy val resultUpper = connection.getDb.get(index.trim.toUpperCase.getBytes())
+    lazy val resultExact = connection.getDb.get(index.trim.getBytes())
 
-  def lookupIndex(index: String): Option[Array[A]] = {
+    if (resultExact != null)
+      Some(fromBytes(resultExact))
+    else if (caseSensitiveIndex && resultLower != null)
+      Some(fromBytes(resultLower))
+    else if (caseSensitiveIndex && resultExact != null)
+      Some(fromBytes(resultExact))
+    else if (caseSensitiveIndex && resultUpper != null)
+      Some(fromBytes(resultUpper))
+    else
+      None
+  }
+
+  def lookup(index: String): Option[Array[A]] = {
     synchronized {
-      lru.getOrElseUpdate(index, customizedLookup(index))
+      lru.getOrElseUpdate(index, lookupByIndex(index))
     }
   }
 
   def containsIndex(index: String): Boolean = {
     val wordBytes = index.trim.getBytes()
-    findLocalDb.get(wordBytes) != null ||
-      (findLocalDb.get(index.trim.toLowerCase.getBytes()) != null) ||
-      (findLocalDb.get(index.trim.toUpperCase.getBytes()) != null)
-
+    connection.getDb.get(wordBytes) != null ||
+      (connection.getDb.get(index.trim.toLowerCase.getBytes()) != null) ||
+      (connection.getDb.get(index.trim.toUpperCase.getBytes()) != null)
   }
 
   override def close(): Unit = {
-    super.close()
+    connection.close()
   }
 
 }
