+++ src/test/scala/com/johnsnowlabs/nlp/annotators/ner/crf/NerCrfApproachTestSpec.scala
@@ -1,8 +1,7 @@
 package com.johnsnowlabs.nlp.annotators.ner.crf
 
 import com.johnsnowlabs.nlp._
-import com.johnsnowlabs.nlp.embeddings.EmbeddingsHelper
-import com.johnsnowlabs.storage.StorageHelper
+import com.johnsnowlabs.nlp.embeddings.WordEmbeddingsIndexer
 import org.scalatest.FlatSpec
 
 class NerCrfApproachTestSpec extends FlatSpec {
@@ -68,11 +67,10 @@ class NerCrfApproachTestSpec extends FlatSpec {
 
   it should "correctly handle entities param" in {
 
-    val smallEmbeddings = EmbeddingsHelper.load(
+    WordEmbeddingsIndexer.load(
       "src/test/resources/random_embeddings_dim4.txt",
       spark,
       "TEXT",
-      false,
       "random_embeddings"
     )
     val restrictedModel = new NerCrfModel()
